make all files and folders auto generate

+++ server.js
@@ -35,41 +35,13 @@ require("./routes/forums/forums")(app)
 require("./routes/forums/post")(app)
 require("./routes/forums/create")(app)
 
-try {
-    fs.mkdirSync("./data")
-    utils.writeJson("./data/users.json", {})
-    utils.writeJson("./data/forums.json", {
-        General: {
-            description: "General channels",
-            channels: {
-                General1: {
-                    description: "General chat channel 1",
-                    private: false,
-                    lastpost: 1,
-                    posts: [
-                        {
-                            title: "Hello World",
-                            timestamp: 1620615527544,
-                            content: "Hello world! This is the test post.",
-                            author: "rhoneeiler@gmail.com",
-                            comments: [
-                                {
-                                    timestamp: 1620615527544,
-                                    content: "Hello world! This is the test comment.",
-                                    author: "rhoneeiler@gmail.com"
-                                }
-                            ]
-                        }
-                    ]
-                }
-            }
-        }
-    })
-} catch (e) {}
-
-try {
-    fs.mkdirSync("./pages")
-} catch (e) {}
+if (!fs.existsSync("./data/")) fs.mkdirSync("./data")
+if (!fs.existsSync("./data/users.json")) utils.writeJson("./data/users.json", {})
+if (!fs.existsSync("./data/forums.json")) utils.writeJson("./data/forums.json", {General: {description: "General channels",channels: 
+{General1: {description: "General chat channel 1",private: false,lastpost: 1,posts: [{title: "Hello World",timestamp: 1620615527544,
+content: "Hello world! This is the test post.",author: "rhoneeiler@gmail.com",comments: [{timestamp: 1620615527544,
+content: "Hello world! This is the test comment.",author: "rhoneeiler@gmail.com"}]}]}}}})
+if (!fs.existsSync("./pages/")) fs.mkdirSync("./pages")
 
 app.get("*", (req, res) => {
     var session = req.session;
